requirements.txt
litellm==1.76.0
playwright==1.50.0
python-dotenv==1.1.1
